I've got typescript and expo and react native working together thanks to this repository https://github.com/janaagaard75/expo-and-typescript
