+++ app/App.js
@@ -1,23 +1,3 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
+import App from './src/App'
 
-export default class App extends React.Component {
-  render () {
-    return (
-      <View style={styles.container}>
-        <Text>Open up App.js to start working on your app!</Text>
-      </View>
-    )
-  }
-}
-
-const defaultBackgroundColor = '#FFF'
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: defaultBackgroundColor,
-    alignItems: 'center',
-    justifyContent: 'center'
-  }
-})
+export default App
