+++ app/package.json
@@ -13,9 +13,14 @@
   "dependencies": {
     "expo": "^30.0.1",
     "react": "16.3.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz"
+    "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz",
+    "tslib": "1.9.3"
   },
   "devDependencies": {
+    "@types/expo": "27.0.13",
+    "@types/expo__vector-icons": "6.2.3",
+    "@types/react": "16.4.16",
+    "@types/react-native": "0.57.4",
     "babel-eslint": "^10.0.1",
     "eslint": "^5.6.1",
     "eslint-config-standard": "^12.0.0",
@@ -24,6 +29,8 @@
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-react": "^7.11.1",
     "eslint-plugin-react-native": "^3.3.0",
-    "eslint-plugin-standard": "^4.0.0"
+    "eslint-plugin-standard": "^4.0.0",
+    "react-native-typescript-transformer": "1.2.10",
+    "typescript": "3.1.2"
   }
 }
